+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "2048 scores".
 - Tests for "Wine from Kalbodia - episode 1".
 - Tests for "2.5D maze".
+- Tests for "Minimal number of swaps".
 
 ## [1.16.0] - 2022-12-31
 ### Added